backend/requirements.txt
absl-py==2.4.0
alembic
amqp==5.3.1
astunparse==1.6.3
async-timeout==5.0.1
bcrypt==5.0.0
billiard==4.2.4
black==23.7.0
blinker==1.9.0
celery==5.3.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.7
dnspython==2.8.0
email-validator==2.0.0
exceptiongroup
flake8==6.1.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
flatbuffers==25.12.19
gast==0.4.0
google-auth==2.48.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.3.2
grpcio==1.74.0
gunicorn==21.2.0
h5py==3.15.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.3.2
keras==2.13.1
kombu==5.6.2
libclang==18.1.1
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
ml_dtypes==0.5.4
mypy_extensions==1.1.0
namex==0.1.0
numpy==1.24.3
oauthlib==3.3.1
opt_einsum==3.4.0
optree
packaging==26.0
pandas==2.0.3
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.2
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==4.25.8
psycopg2-binary==2.9.7
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==3.0
pyflakes==3.1.0
Pygments==2.19.2
PyJWT==2.11.0
pymongo==4.4.1
pytest==7.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
redis==4.6.0
requests==2.31.0
requests-oauthlib==2.0.0
rich==14.3.3
reportlab==4.0.4
rsa==4.9.1
scikit-learn==1.3.0
scipy==1.15.3
six==1.17.0
SQLAlchemy
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
threadpoolctl==3.6.0
tomli==2.4.0
typing_extensions
tzdata==2025.3
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
Werkzeug==3.1.6
wrapt==2.1.1
